Make tests 5x faster
