+++ soroban-sdk/Cargo.toml
@@ -12,6 +12,7 @@ rust-version = "1.64"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
+doctest = false
 
 [dependencies]
 soroban-sdk-macros = { version = "0.0.4" }
@@ -25,8 +26,10 @@ soroban-env-host = { version = "0.0.5", features = ["vm", "hostfn_log_fmt_values
 ed25519-dalek = { version = "1.0.1", optional = true }
 
 [dev-dependencies]
+soroban-env-host = { version = "0.0.5", features = ["vm", "hostfn_log_fmt_values", "testutils"] }
 stellar-xdr = { version = "0.0.2", features = ["next", "std"] }
 soroban-spec = "0.0.4"
+ed25519-dalek = "1.0.1"
 trybuild = "1.0.63"
 hex = "0.4.3"
