+++ soroban-sdk/src/tests.rs
@@ -1,0 +1,12 @@
+#![cfg(test)]
+
+mod assert_panic;
+mod contract_add_i32;
+mod contract_invoke;
+mod contract_udt_enum;
+mod contract_udt_struct;
+mod contract_udt_struct_tuple;
+mod contractfile_with_sha256;
+mod contractimport;
+mod contractimport_with_sha256;
+mod trybuild;
