+++ soroban-sdk/src/tests/contract_udt_struct.rs
@@ -1,8 +1,6 @@
-#![cfg(feature = "testutils")]
-
+use crate as soroban_sdk;
 use soroban_sdk::{
-    contractimpl, contracttype, vec, BytesN, ConversionError, Env, IntoVal, RawVal, TryFromVal,
-    TryIntoVal, Vec,
+    contractimpl, contracttype, map, symbol, BytesN, ConversionError, Env, TryFromVal,
 };
 use stellar_xdr::{
     ReadXdr, ScSpecEntry, ScSpecFunctionInputV0, ScSpecFunctionV0, ScSpecTypeDef, ScSpecTypeTuple,
@@ -10,8 +8,11 @@ use stellar_xdr::{
 };
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
-#[contracttype]
-pub struct Udt(pub i32, pub i32);
+#[contracttype(crate_path = "crate")]
+pub struct Udt {
+    pub a: i32,
+    pub b: i32,
+}
 
 pub struct Contract;
 
@@ -23,22 +24,13 @@ impl Contract {
 }
 
 #[test]
-fn test_conversion() {
-    let env = Env::default();
-    let a = Udt(5, 7);
-    let r: RawVal = a.into_val(&env);
-    let v: Vec<i32> = r.try_into_val(&env).unwrap();
-    assert_eq!(v, vec![&env, 5, 7]);
-}
-
-#[test]
 fn test_functional() {
     let env = Env::default();
     let contract_id = BytesN::from_array(&env, &[0; 32]);
     env.register_contract(&contract_id, Contract);
 
-    let a = Udt(5, 7);
-    let b = Udt(10, 14);
+    let a = Udt { a: 5, b: 7 };
+    let b = Udt { a: 10, b: 14 };
     let c = ContractClient::new(&env, &contract_id).add(&a, &b);
     assert_eq!(c, (a, b));
 }
@@ -47,23 +39,23 @@ fn test_functional() {
 fn test_error_on_partial_decode() {
     let env = Env::default();
 
-    // Success case, a vec will decode to a Udt.
-    let map = vec![&env, 5, 7].to_raw();
+    // Success case, a map will decode to a Udt if the symbol keys match the
+    // fields.
+    let map = map![&env, (symbol!("a"), 5), (symbol!("b"), 7)].to_raw();
     let udt = Udt::try_from_val(&env, map);
-    assert_eq!(udt, Ok(Udt(5, 7)));
+    assert_eq!(udt, Ok(Udt { a: 5, b: 7 }));
 
-    // If a struct has 2 fields, and a vec is decoded into it where the vec has
-    // 2 elements, it is an error. It is an error because all fields must be
-    // assigned values.
-    let map = vec![&env, 5, 7, 9].to_raw();
-    let udt = Udt::try_from_val(&env, map);
-    assert_eq!(udt, Err(ConversionError));
-
-    // If a struct has 2 fields, and a vec is decoded into it where the vec has
-    // 3 elements, it is an error. It is an error because decoding and encoding
-    // will not round trip the data, and therefore partial decoding is
-    // relatively difficult to use safely.
-    let map = vec![&env, 5, 7, 9].to_raw();
+    // If a struct has fields a, b, and a map is decoded into it where the map
+    // has fields a, b, and c, it is an error. It is an error because decoding
+    // and encoding will not round trip the data, and therefore partial decoding
+    // is relatively difficult to use safely.
+    let map = map![
+        &env,
+        (symbol!("a"), 5),
+        (symbol!("b"), 7),
+        (symbol!("c"), 9)
+    ]
+    .to_raw();
     let udt = Udt::try_from_val(&env, map);
     assert_eq!(udt, Err(ConversionError));
 }
@@ -73,7 +65,7 @@ fn test_spec() {
     let entries = ScSpecEntry::from_xdr(__SPEC_XDR_ADD).unwrap();
     let expect = ScSpecEntry::FunctionV0(ScSpecFunctionV0 {
         name: "add".try_into().unwrap(),
-        inputs: std::vec![
+        inputs: vec![
             ScSpecFunctionInputV0 {
                 name: "a".try_into().unwrap(),
                 type_: ScSpecTypeDef::Udt(ScSpecTypeUdt {
@@ -89,8 +81,8 @@ fn test_spec() {
         ]
         .try_into()
         .unwrap(),
-        outputs: std::vec![ScSpecTypeDef::Tuple(Box::new(ScSpecTypeTuple {
-            value_types: std::vec![
+        outputs: vec![ScSpecTypeDef::Tuple(Box::new(ScSpecTypeTuple {
+            value_types: vec![
                 ScSpecTypeDef::Udt(ScSpecTypeUdt {
                     name: "Udt".try_into().unwrap(),
                 }),
