+++ soroban-sdk/src/tests/contractfile_with_sha256.rs
@@ -1,3 +1,4 @@
+use crate as soroban_sdk;
 pub const WASM: &[u8] = soroban_sdk::contractfile!(
     file = "../target/wasm32-unknown-unknown/release/example_add_i32.wasm",
     sha256 = "f0554f39a5ea3fe414d11b52831ea9a5a65e62a171b8c18f2bd51b2365cbb242",
