+++ soroban-sdk/src/tests/trybuild/contractfile_with_sha256_verify_fail.rs
