+++ soroban-sdk/src/tests/trybuild/contractfile_with_sha256_verify_fail.stderr
@@ -1,11 +1,11 @@
 error: sha256 does not match, expected: f0554f39a5ea3fe414d11b52831ea9a5a65e62a171b8c18f2bd51b2365cbb242
- --> tests/trybuild/contractfile_with_sha256_verify_fail.rs:3:5
+ --> src/tests/trybuild/contractfile_with_sha256_verify_fail.rs:3:5
   |
 3 |     sha256 = "0000000000000000000000000000000000000000000000000000000000000000",
   |     ^^^^^^
 
 error[E0601]: `main` function not found in crate `$CRATE`
- --> tests/trybuild/contractfile_with_sha256_verify_fail.rs:4:3
+ --> src/tests/trybuild/contractfile_with_sha256_verify_fail.rs:4:3
   |
 4 | );
-  |   ^ consider adding a `main` function to `$DIR/tests/trybuild/contractfile_with_sha256_verify_fail.rs`
+  |   ^ consider adding a `main` function to `$DIR/src/tests/trybuild/contractfile_with_sha256_verify_fail.rs`
