+++ soroban-sdk/src/tests/trybuild/contractimpl_fn_name_too_long.stderr
@@ -1,5 +1,5 @@
 error: contract function name too long, max length 10 characters
- --> tests/trybuild/contractimpl_fn_name_too_long.rs:7:12
+ --> src/tests/trybuild/contractimpl_fn_name_too_long.rs:7:12
   |
 7 |     pub fn pay_to_destination() {
   |            ^^^^^^^^^^^^^^^^^^
