+++ soroban-sdk/src/tests/trybuild/contractimpl_udt_generics.rs
