+++ soroban-sdk/src/tests/trybuild/contractimpl_udt_generics.stderr
@@ -1,11 +1,11 @@
 error: generics unsupported on user-defined types in contract functions
- --> tests/trybuild/contractimpl_udt_generics.rs:9:25
+ --> src/tests/trybuild/contractimpl_udt_generics.rs:9:25
   |
 9 |     pub fn add(a: MyType<i32>, b: MyType<i32>) -> i8 {
   |                         ^
 
 error: generics unsupported on user-defined types in contract functions
- --> tests/trybuild/contractimpl_udt_generics.rs:9:41
+ --> src/tests/trybuild/contractimpl_udt_generics.rs:9:41
   |
 9 |     pub fn add(a: MyType<i32>, b: MyType<i32>) -> i8 {
   |                                         ^
