+++ soroban-sdk/src/tests/trybuild/contractimport_with_sha256_verify_fail.rs
