+++ soroban-sdk/src/tests/trybuild/contracttype_enum_field_name_too_long.rs
