+++ soroban-sdk/src/tests/trybuild/contracttype_struct_field_name_too_long.rs
