+++ soroban-sdk/src/tests/trybuild/contracttype_struct_field_name_too_long.stderr
@@ -1,5 +1,5 @@
 error: struct field name symbol too long: length 11, max 10
- --> tests/trybuild/contracttype_struct_field_name_too_long.rs:6:9
+ --> src/tests/trybuild/contracttype_struct_field_name_too_long.rs:6:9
   |
 6 |     pub f2istoolong: bool,
   |         ^^^^^^^^^^^
