+++ soroban-sdk/src/testutils.rs
@@ -1,4 +1,4 @@
-#![cfg(feature = "testutils")]
+#![cfg(any(test, feature = "testutils"))]
 #![cfg_attr(feature = "docs", doc(cfg(feature = "testutils")))]
 
 //! Utilities intended for use when testing.
